fix hover over line issue
Switch to class based modifier to allow initial state to be separated
from updates.

Don't tear down the svg group when the styling of the paths change.

If everything is torn down it causes the element to be appended after
the mouse over box when the colour and stroke of the hovered line is
updated. This in turn causes the mouse over box to loose focus and the
colour and stroke of the line to be reverted.